Spring security end point

+++ src/main/java/net/engineeringdigest/journalApp/Service/JournalEntryService.java
@@ -24,7 +24,7 @@ public class JournalEntryService {
         try{ User user = userService.findByUsername(username);
             JournalEntry saved = journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
-            userService.saveEntry(user);
+            userService.saveUser(user);
         } catch (Exception e){
             System.out.println(e);
             throw new RuntimeException("An error occured while saving entries",e);
@@ -41,13 +41,25 @@ public class JournalEntryService {
     public Optional<JournalEntry> findbyId(ObjectId id){
     return journalEntryRepository.findById(id);
     }
-    public void deletebyId(ObjectId id, String username){
 
+    @Transactional
+    public boolean deletebyId(ObjectId id, String username){
+        boolean removed = false;
+    try {
         User user = userService.findByUsername(username);
-        user.getJournalEntries().removeIf(x->x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+         removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+        if (removed) {
+            userService.saveUser(user);
+            journalEntryRepository.deleteById(id);
+        }
+    } catch(Exception e){
+        System.out.println(e);
+        throw new RuntimeException("An error occured while deleting the entry" , e);
 
-        journalEntryRepository.deleteById(id);
+        }
+
+        return removed;
+
     }
+
 }
